backend/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
autoflake==1.4
Automat==20.2.0
backports.entry-points-selectable==1.1.0
black==21.7b0
cffi==1.14.6
cfgv==3.3.0
channels==3.0.4
channels-redis==3.3.0
click==8.0.1
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
distlib==0.3.2
Django==3.2.6
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.2
flake8-bugbear==21.4.3
future==0.18.2
hiredis==2.0.0
hyperlink==21.0.0
identify==2.2.13
idna==3.2
incremental==21.3.0
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.2.0
pre-commit==2.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyOpenSSL==20.0.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
virtualenv==20.7.2
zope.interface==5.4.0
